cherry-pick: allow filtering by source directory
Instead of formatting a patch on the entire source repo, the -S argument
allows picking a patch confined to a specific subdirectory and applying
it to the root of the destination repo. This allows cherry-picking
patches from a vendored third_party/ into the upstream repo without
pulling in changes from the other parts of the monorepo.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>