cherry-pick: add trailer customization
The '--trailer' argument is for automating cherry-pick workflows by
adding RFC822-style trailers to the end of the commit message before
applying the patches. These trailers are applied with
git-interpret-trailers(1) and follow the same format.

There is a lot of customization available throught the trailer.* config
variables and command-line arguments to git-interpret-trailers(1). The
command-line options are not going to be exposed through
git-third-party, but the config file will still work for modifying
trailer behavior.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>