clone: vendor a squashed tree of a third-party repo
Initial checkin of the clone command that allows vendoring a third-party
repo into the monorepo. This initial revision requires the --squash
argument, which copies the specified branch into the toplevel git
directory, removing any git history. The code is not committed in the
monorepo.

The code is vendored in the directory passed using the -d argument,
relative to the toplevel git directory.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>